fix HEAD request incompatible
HEAD requests to S3 respond with the exact same header as a GET,
swift3 responds with an incorrect Content-Type and Content-Length. A
quick hack to fix this issue is below, but should probably be replaced
with something involving DiskFile and such from Swift's API, because
of innecessary overhead involved in reading in the file.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>