Fix listing bucket when key contains special URL charcters
When a key contains a special URL char (ex: %20) the key
name is unquote when listed in a bucket.
This result is having, for example "my name" and "my%20name"
be listed the same way allthough they are two different objects.

This patch remove unquote call on the object name
to stick with AWS S3 behavior.

Change-Id: Icc104876a36f91ade738b998174d2def1b86837c