Add setup.cfg and pbr support.
This patch fix a bug when installing a swift3 on
devstack.

Because devstack uses pbr/requirements.txt to
sync global requirements, we need to add setup.cfg and
pbr support to swift3 so it can be installed on devstack.

This also allow to take advantage of the pbr facilities
(version number support, etc)

Change-Id: Id0179fad179efac5ba85bef7838365a4d304d196