Use lxml for processing XML
The lxml XML toolkit is a binding for the C libraries libxml2 and libxslt.  It
is compatible with ElementTree API, which is more pythonic than the DOM
interface.  In addition, lxml has attractive features like XML validation and
namespace mapping.

Replacing the current DOM interface with the lxml API will simplify Swift3 code
a lot and improve maintainability, I believe.

Change-Id: Ie2291f180421559ed3320a173d2f7eea81f459d9