request: forbid request.body
swob.Request.body is not secure against malicious input.  It consumes too much
memory when the request body is excessively large.

This patch forbids insecure body() and introduces xml(), which limits the
maximum length of body, instead.  The maximum values this patch adds are same
ones as AWS S3 uses.

Change-Id: I8f3a14fc442338d77080951458795991d56c6cfa