etree: add support for non-ascii string
Swift3 doesn't support a non-ascii string because the lxml library handles it
as unicode, which Swift doesn't expect.  A simple but silly way to fix the
problem is to call encode/decode every time we access lxml.etree._Element.text.
E.g.,

  if isinstance(value, unicode):
     value = elem.text.encode('utf8')
  else:
     value = elem.text

This is bad for maintainability.

The better way would be to override the text method in lxml.etree._Element, but
we cannot do it unfortunately because the class is not pure python.

This patch introduces a wrapper class for lxml.etree._Element, and override the
text method of it.  The swift3.etree module would become a bit tricky, but we
don't need to change other code at all.

Change-Id: Ifcffcfdb2acb6e1d088e803c0432d52dae3d44ca