Fix Multi Delete broken with non-ascii keys
When non-ascii keys are used in Multi Delete xml,
swift3 can not handle them because swift3 does not
expect the keys are given by unicode string.

This patch introduces get_valid_utf8_str() function from
swift. It serves a correct str for a path even though a
src string is in either case of str and unicode.

Change-Id: I4b0f23b1c011a982c6250ca03eb403e03baa82ea