Fix response on object copy requests.
In swift3/controllers/obj.py, the PUT handler checkes if a copy response is required.
However, it checks if a Swift copy header is present, but the request it inspects is
the original AWS request, which will contain the AWS equivalent of that header. This
patch fixes this.

Also, the LastModified header was missing from the XML report, this patch readds it.

Change-Id: Ib076ff4dd16edb4bddd1051c67f5ee7749b86740
Closes-Bug: #1378060