Fix Canned ACL Error with Empty Body
Current acl controller do a double check both ACL header and
ACL body (xml).

However, if ACL body is not given but ACL header (X-AMZ-ACL) is
given as a canned ACL, current implementation always causes ACL
Error in parsing XML since the request body is empty even though
it should be handled as a correct request on acl controller.
In addition, if both of them are given, we have to UnexpectedContent
Exception according to S3 behavior.

This patch fixes to try to parse the request body only when
the request body is not empty. This pach also fixes to raise
if both ACL header and xml are given.

Change-Id: Ie23d5e3d94c47e0411197afd22f0b31db2ac45d1
Closes-Bug: 1382290