Draft: s3acl decorator
Current swift3 has a couple of ways for acl handling.
To simplify test code, this introduces s3acl decorator
which suggests which way the test method will check.

In default, swift3 do the native acl testing. If needed,
we can add args to decorator as follows:

No decorator -> test only native acl
@s3acl -> test both of native acl and s3acl
@s3acl(s3acl_only=True) -> test only s3acl

This is still a draft version and most of tests should
be fixed.

Change-Id: I17a423b5c655a670a4304d47ff927df28ac0a37f