Fix max-parts, part-number-marker, and encoding-type in queries of List Parts
I've fixed ToDo in qeuries of  List Part.

encoding-type: Requests Amazon S3 to encode the response and specifies the encoding method to use.
max-parts: Sets the maximum number of parts to return in the response body.
part-number-marker: Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

For details, refer to the following.
http://docs.aws.amazon.com/AmazonS3/latest/API/mpUploadListParts.html

Change-Id: I1a3b3a0aba188fcb928e89e78ad8859dbecca2e8