Add config paramater `pipeline_check` to enable/disable pipeline check
`pipeline_check` is an option to disable the exception when the order of
middlewares in pipeline is incorrect. The pipeline check is enabled in
default.

Disable pipeline check that makes swift supported multiple authentications.
While it's enabled with keystoneauth in pipeline, swift3 force to check
s3token, authtoken must be in pipeline. And that will make all swift3
requests authenticate with keystoneauth only, can't bypass keystoneauth to
other 3party authenticate middlewares. So, turn pipeline_check off is a
way to do that.

Usage:

    [filter:swift3]
    pipeline_check = False

Change-Id: Ic3f3fc8454f13be587a4903bae02274aa999324c
Closes-Bug: #1407594