Assemble similar stuffs into a validation method
Currently, controllers have similar stuffs to validate
query parameter either to limit the listing result or
to make a marker starts with next to it.

This patch intoduces a validation method 'get_validated_param'
for those stuffs to get a pre-validated value with some parameters
(e.g. default, limit) to keep the code clean.

Change-Id: Ia5e939ee342bf4fb8533f1aa97419f5aad9b7fe1