Fix Swift3 to skip S3 authorization after initial authentication
Fix Swift3 not sent to again header('Authorization') after the initial
authentication completion, if it is used keystone.
Current Swift3 sends again this header after authentication, when s3_acl
is effective and requested GET or HEAD Object.
In the above case, 'swift.authorize_override' is not effective and Swift3
returns 403.

Change-Id: Id81e393d51b389610d9fa470f307f61e846a78a3