Fix the maximum value of part-number-marker for List Parts
Fix to be able to specify part-number-marker the range of 0-2147483647.
(2147483647 means the maximum value of integer.)

PartNumber for Upload Part been able to specified the range of 1-10000 in S3.
But the range of part-number-marker is not consistent with the above-mentioned.
If you specify part-number-marker the range of 10000-2147483647, 200 OK
and body that does not containe Part tag will be returned in S3.

Currently part-number-marker been able to specify up to the maximum value of
max-parts in Swift3.
However, that dose not match the specifications of S3.

Change-Id: I345f378d4da1cf9ada42147e56b55cb915d53582