Fix get_validated_param and APIs related to this method
Fix get_validated_param and the response of the APIs related to this
method to compatible with S3 specifications.

Combinations of target parameters and tag corresponding to each
them are following.
 - max-keys: <MaxKeys/>
 - max-uploads: <MaxUploads/>
 - max-parts: <MaxParts/>
 - part-number-marker: <PartNumberMarker/>

When the value of above parameters are not in the range of positive
32bit integer, S3 returns 400.

When the value of above parameters are in the range of 1001-2147483647
(2147483647 means max 32bit integer), handling of S3 depending on the
type of the parameter is different.
See below for handling of each parameter when specified the above range.

1. max-keys for GET Bucket
 - the response status: 200
 - the value of <MaxKeys/>: specified value
 - the maximum number of objects that can be acquired: 1000

2. max-uploads for List Multipart Uploads
 - the response status: 200
 - the value of <MaxUploads/>: 1000
 - the maximum number of uploads that can be acquired: 1000

3. max-parts for List Parts
 - the response status: 200
 - the value of <MaxParts/>: 1000
 - the maximum number of parts that can be acquired: 1000

4. part-number-marker for List Parts
  - the response status: 200
  - the value of <PartNumberMarker/>: specified value
  - acquisition start position of part number: specified value

Change-Id: I8bf5b2816238b05469ceef5758d793f2fa4b3225