Keystone_token is no longer needed
Swift3 uses 2 types of tokens when using keystone auth. One is a token
generated from upload information as a canonical string. This token is used
with "Authorization" header. The other is a token retreived from a keystone
server as a pre-autheticated token. This token is used without "Authorization"
header to clarify it as a pre-authenticated request.

We've already have a code to delete that "Authorization" header when the
pre-authenticated request succeeded so that swift3 doesn't need to keep
the keystone_token variable in the Request class.

Change-Id: I27dfff910a0d5517d00809bb65041c3bb5dd35ca