Fix the version of Swift that required to install Swift3
Fix the version of the Swift described in requirements.txt from 1.8.0
to 2.0.0.

In between updates from 1.13.1 to 2.0.0, the process for pipeline of
Swift was changed.
Therefore, the above version 2.0.0 is not installed, starting of
proxy-server with Swift3 fails.

For more information about the change of Swift, check the following
commit ID.
- 4ce9b252fd44dae74b47d5b5c48e9403b5d36a4a

Change-Id: I793b0f325ea34f208784d60f2b038e75f6d239b9