Fix the version of Swift that required to install Swift3
Fix the version of the Swift described in requirements.txt from 1.8.0
to 2.1.0.

In between updates from 1.8.0 to 2.0.0, the process for pipeline of
Swift was changed.
And object sysmeta metadata was supported from 2.1.0.

Therefore, 2.1.0 or later is not installed, starting of
proxy-server with Swift3 fails and it is impossible to use s3_acl.

For more information about the changes of Swift, check the following
commit IDs.
- 4ce9b252fd44dae74b47d5b5c48e9403b5d36a4a
- 4286f36a60446e322f23a329416be3ed67db1149

Change-Id: I793b0f325ea34f208784d60f2b038e75f6d239b9