Support delimiter for List Multipart Uploads
Support delimiter for List Multipart Uploads.

In order to associate object name and upload ID, slash(/) is used
in Swift3.
So, sent request to Swift with delimiter specified a slash(/), Swift
does not return expected response.
ex) Swift has following objects.
     - obj/uploadId
     - obj/uploadId/1
     - obj/uploadId/2
     - dir/obj/uploadId
     - dir/obj/uploadId/1
     - dir/obj/uploadId/2
    Specified delimiter is slash(/), Swift returns following.
     [{'subdir': 'obj/', ...},
      {'subdir': 'dir/', ...}]
    But Swift3 expects following.
     [{'name': 'obj/uploadId', ...},
      {'name': 'obj/uploadId/1', ...},
      {'name': 'obj/uploadId/2', ...},
      {'subdir': 'dir/', ...}]

That is why, it is not to sent delimiter to Swift, all handling for
delimiter is performed in Swift3.

Change-Id: Ic04662c35a9bc409132e6e8175eb8159690b37a4