Add mktime function to test.functional.utils
email.utils.parsedate and time.mktime is really useful
for standard timestamp translation because we don't have
to make a handmade format translation logic.

However, it tends to make a long formula because everytime
we have to call both parsedate and mktime to translate from
RFC2822 format. In addition, unfortunately S3 uses also
different timestamp formats from RFC2822 in the response xml.

To get an easy way for the translation work, this patch
add the mktime function to translate from any format supported
by S3 to a float in epoch time.

TODO:
Currently mktime in this patch doesn't support the format of
list objects last-modified timestamp. We should do it on the
later patch.

Change-Id: I51888bc625773f7673fe01e7bdc6636cf3709901