Fix non-exist bucket bugs on multi-upload
List parts, Abort upload and Complete upload expects
404 NoSuchBucket when the target bucket doesn't exist.

This patch fixes the responses by making swift3 to check
the bucket exisitence.

This patch also includes minor changes on the unit tests side.
That is because we don't want a lot of duplicated patch decoratores
on the top of the test method and almost of those are just normal
status as 204 NoContent against to HEAD container so this patch adds
it as a default. If get_container_info is not called on the test,
the patching will be just ignored.

If you want, we can rewrite the statement easily with nested decorator
like as:

@s3acl
@patch(xxx)
def test_xxxx(self):

Change-Id: I1b29ad98a4424faef5598b0cc1cdf1f6dee4dd99
Closes-Bug: #1440664
Closes-Bug: #1440667
Closes-Bug: #1440668