Add more sanity for HEAD requests
According to HTTP specification, a HTTP server *must not* response
body at HEAD request. For now, Swift3 supports the correct behavior
but it depends on swift.common.swob.Response class and it *might* be
confusable because the code is maintained in another place.

This patch adds unit tests to check both the Response class behavior and
that Swift3 is now following the HTTP specifications.

Change-Id: Ib7e906cb785e84716dbb9ea9d311bbf6a1b5fc2e
Closes-bug: #1464450