Update version of keystone used for functests
Previously, the gate tests used a version of keystone that didn't have a
cap on oslo.config. This led to gate checks failing because they pulled
in a version of oslo.config that broke backward compatibility. Even if
we updated within icehouse, the last oslo.messaging does not include a
cap on oslo.utils, which *also* broke backwards compatibility.

Now, we'll install the most recently released version of keystone,
updating some testing config along the way.

Change-Id: Id357975413094bab751c5b8549d9201e9232af7f