Update version of keystone used
Previously, the gate tests used a version of keystone that didn't have a
cap on oslo.config. This lead to gate checks failing because they pulled
in a version of oslo.config that broke backward compatibility.

Now, we'll install the lowest version of icehouse keystone that includes
a version cap.

Change-Id: Id357975413094bab751c5b8549d9201e9232af7f