Add object copy to self checks
This patch adds checks when copying an object to itself.

Currently, Swift3 allows the user to copy an object to itself, without an error
message.

However, on S3, an InvalidRequest response occurs when no attributes of the
object (such as metadata, storage class, website redirect location, or
encryption) change.  Currently, only metadata is supported, however, if other
attributes are supported in the future, more checks will be needed.

Note: S3 appears to accept the X-Amz-Copy-Source header both with and without a
leading '/'

Change-Id: Ifc3e513321953f1200383302ef38bbd85fa86b92