Translate Cache-Control and Expires headers for responses
Per AWS's docs [1], Cache-Control and Expires may be set on upload. On
download, the same headers would then be included in the response.
Previously, these would not be included in Swift3 responses; now they
will be.

Note that setting the headers already worked in Swift3, but requires
that you update the allowed_headers option in the [app:object-server]
section of object-server.conf. The conf used for functional tests has
been so updated.

[1] http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectPUT.html#RESTObjectPUT-requests-request-headers

Change-Id: I22001c6fd14033a9f13c36a3e05fdc678c75654f