Properly handle custom metadata upon an object COPY operation
Previously, the 'x-amz-metadata-directive' header was ignored except to
ensure that it had a valid value if it existed. In all cases any
metadata specified was applied to the copied object, while
non-conflicting metadata remained.

This patch fixes this behaviour.

Now, if the 'x-amz-metadata-directive' header is set to 'REPLACE'
during a copy operation, the swift3 middleware sets the
'x-fresh-metadata' header to 'True' to replace valid metadata values.

If the 'x-amz-metadata-directive' header is set to 'COPY' or if it is
omited during a copy operation, then the swift3 middleware removes all
metadata (custom or not) from the request to prevent it from being
changed.

Content-Type can never be set on an S3 copy operation, even if the
metadata directive is set to 'REPLACE', so it is specifically filtered
out on copy.

Also fixed an issue where items were being removed and added to a
dictionary while iterating, which could have resulted in some items
being skipped.

Change-Id: I30687f4e2ff04bb4091f54ebbeedf91b9052150f
Closes-Bug: #1433875