Ensure object copy XML response is consistent with S3
The LastModified tag now properly reflects the timestamp of the newly copied
object, rather than the original object.

The LastModified timestamps now includes ".000Z" at the end to be consistent
with S3 responses.

The ETag tag now uses '&quot;' instead of '"' to surround the ETag value to be
consistent with the S3 response.

Change-Id: Iffe70cb49701142a09c75b7cc33aaaea4e846144
Closes-Bug: #1496170