Accept URL-encoded x-amz-copy-source headers
Previously, these were not decoded, leading to a double-encoding when
translating to a X-Copy-From header. This would result in 404s when the
source object had, for example, spaces in its name.

It should be noted that per
http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectCOPY.html,
"[The x-amz-copy-source] string must be URL-encoded." Empirically,
however, AWS also accepts unencoded sources (hence the
quote(unquote(...)) shenanigans).

Change-Id: I20ce0966198cf9d0f59692591bea9381ea57da51