Ignore Content-Type from client when deleting multi-part objects
The aws-sdk-java library was observed sending DELETE requests with a
Content-Type of "application/x-www-form-urlencoded; charset=utf-8".
If we allow that through, SLO will pass it along to the bulk-deleter,
which will respond with:

    HTTP/1.1 200 OK
    [various headers...]

    Number Deleted: 0
    Number Not Found: 0
    Response Body:
    Response Status: 406 Not Acceptable
    Errors:

...and nothing will have been deleted. Worse, since the real error code
is buried in the body, it's difficult to detect the underlying cause.

Now, we'll just drop any content-type we receive from the client.
Another patch will be submitted to do something similar for SLO, but
this will fix multi-part deletes for older versions of Swift.

Related-Change: I791c7bda1d9df830d8dacd3783c2393db5a9ac09
Change-Id: I334726c61990fdbc984855d8ce562774c477cce9