WIP: Refacor sig v4 patch
Currently we have a lot of Authentication parsing accross in
the S3Request class and it makes us complicated to know what's
statement there.

This patch add a new decorator function to attach sig v4 functionality
to existing classes.

I am trying to gather/split the process to be chunks which can
be maitained easyliy but this patch is still in working in progress.

Change-Id: I9b041c84e9144fdf83e1161adaea62988f8f3b78