WIP: Refacor sig v4 patch
Currently we have a lot of Authentication parsing accross in
the S3Request class and it makes us complicated to know what's
statement there.

This patch add a new decorator function to attach sig v4 functionality
to existing classes.

I am trying to gather/split the process to be chunks which can
be maitained easyliy but this patch is still in working in progress.

What I did in this patch:

- Move v4 related code into signature_v4 function which is a class
  decorator to attach some methods for authentication algorithms (e.g.
  string_to_sign)

- Make sigv4 classes with the decorator

- Ommit path normalization code for relative path support because I
  couldn't get if S3 supports such a syntax (once I attmpted it via boto
  client, nothing worked well as I expected. i.e. all path looked to be
  handled as it is.)

Change-Id: I9b041c84e9144fdf83e1161adaea62988f8f3b78