Include '-' in multipart ETags
Multipart uploads in AWS (seem to) have ETags like:

   '"' + MD5_hex(MD5(part1) + ... + MD5(partN)) + '-' + N + '"'

On the other hand, Swift SLOs have Etags like:

   MD5_hex(MD5_hex(part1) + ... + MD5_hex(partN))

(In both examples, MD5 gets the raw 16-byte digest while MD5_hex
gets the 32-byte hex-encoded digest.)

Some clients (such as aws-sdk-java) use the presence of a dash
to decide whether to perform client-side validation of downloads.

Other clients (like s3cmd) use the presence of a dash *in bucket
listings* to decide whether or not to MD5 files locally.

Now we include a dash as well, to prevent spurious errors like

> Unable to verify integrity of data download.  Client calculated
> content hash didn't match hash calculated by Amazon S3.  The data
> may be corrupt.

For new multipart-uploads via the S3 API, the ETag that is stored will
be calculated in the same way that AWS uses. This ETag will be used in
GET/HEAD responses, bucket listings, and conditional requests via the S3
API; accessing the same object via the Swift API will use the SLO Etag.

New SLOs and pre-existing multipart-uploads may have "-N" appended to
their ETag. Note that this uses the letter "N", as the number of
segments is not known. This modified ETag will be used in key GET/HEAD
responses via the S3 API, while the original ETag will be used for:

 * object responses via the Swift API,
 * container (or bucket) listings, and
 * conditional requests via the Swift API.

Either the modified or original ETag may be used for conditional
requests via the S3 API.

Change-Id: If6f756fb75d926f8b94f52cd4b3b76c261bed61e
Closes-Bug: 1522578