Fix sigv2 presigned urls
We stopped setting the Date header to the Expires query parameter
during the sigv4 work, probably because it looked weird and bad.

Turns out, we actually needed that, or something like it. Added the
Expires parameter to the string to sign as well as some func tests.

Some archived AWS docs may be helpful:

https://web.archive.org/web/20131203082418/http://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html#CreatingASignature

In particular:

> Note also that in StringToSign, the HTTP Date positional element has
> been replaced with Expires.

Change-Id: Id4e6ca01bec57f14bd6775b20f0a90808c2b9a0b
Closes-Bug: 1593866