Add check_signature function to swift3.auth_details
This allows (some) auth middlewares to stop needing to know the details
of S3 request signing. The function takes the user's secret and returns
True if the signatures match, False otherwise.

This offers an easy way for auth middlewares that have user secrets
in-memory (such as swauth and tempauth) to add support for v4 signatures.

Change-Id: Iafb6114c12deb9a40d0f8324611de27b48ed95f6