Introduce auth middleware using account metadata.
Currently, swift3 can only be used with keystone, tempauth (and likely swauth).
Custom auth middlewares must always be adapted in order to be useable
for swift3 (by retrieving the secret key). It is not always required or possible
(e.g. if the auth middleware has no access to the secret key, because it is
already hashed, e.g.) that a custom auth middleware implements the gathering of
the secret of its own.

This commit adds a super light-weight auth middleware, which simply uses
an account metadata entry called s3-secret for authorization. Operators
can just use this middleware instead of writing their own middleware or adapt
their middleware currently in use.

Change-Id: I0908d639f20b9b9fd8cdb5c3d8b44edc6b71c81e